Change minor details
