+++ js/components/HotlineDropdownPicker.js
@@ -3,7 +3,7 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import { View } from 'react-native';
 import DropDownPicker from 'react-native-dropdown-picker';
-
+import { DEFAULT_HOTLINE, HOTLINE_OPTIONS } from '../../data/hotlineOptions';
 import { colors, textStyles } from '../styles';
 
 const HotlineDropdownPicker = ({ setHotlineName, setHotlineNumber }) => {
@@ -11,8 +11,8 @@ const HotlineDropdownPicker = ({ setHotlineName, setHotlineNumber }) => {
     <View style={{ flex: 1, paddingVertical: 20 }}>
       <DropDownPicker
         items={HOTLINE_OPTIONS}
-        defaultValue={DEFAULT_HOTLINE_NUMBER}
-        containerStyle={{ height: 50 }}
+        defaultValue={DEFAULT_HOTLINE.value}
+        containerStyle={{ flexGrow: 0.1 }}
         style={{ backgroundColor: colors.primaryLighter }}
         dropDownStyle={{ backgroundColor: colors.primaryLighter }}
         itemStyle={{ justifyContent: 'flex-start' }}
@@ -36,36 +36,3 @@ HotlineDropdownPicker.propTypes = {
 };
 
 export default HotlineDropdownPicker;
-
-export const HOTLINE_OPTIONS = [
-  {
-    label: 'San Diego Rapid Response Network (San Diego County)',
-    value: '6195360823',
-  },
-  {
-    label:
-      'Boyle Heights Immigrant Rights Network (Los Angeles City - Boyle Heights)',
-    value: '3239225644',
-  },
-  {
-    label:
-      'Southern Central Coast Rapid Response Network (Southern Central Coast Region - Santa Barbara County)',
-    value: '8058708855',
-  },
-  {
-    label: 'Long Beach Community Defense Network (Long Beach, CA)',
-    value: '5622691083',
-  },
-  {
-    label: 'The Koreatown Popular Assembly (Los Angeles City - Koreatown)',
-    value: '3238941504',
-  },
-  { label: 'IC4IJ (San Bernardino / Inland Empire)', value: '9093614588' },
-  {
-    label: 'Orange County Rapid Response Network (Orange County)',
-    value: '6572100157',
-  },
-];
-
-export const DEFAULT_HOTLINE_NAME = HOTLINE_OPTIONS[0].label;
-export const DEFAULT_HOTLINE_NUMBER = HOTLINE_OPTIONS[0].value;
